Default enforceSingleInvocation
Since the introduction of [RN autolink](https://reactnative.dev/docs/linking-libraries-ios#automatic-linking), we don't have the opportunity to instantiate the module ourselves anymore, so it's not possible to pass a custom constructor parameter. Diving into the code, I realized this value is just hardcoded in the source, so I changed it according to the readme, that claims it must be initialized with true to avoid redundant values.